+++ packages/share_plus/share_plus_linux/lib/share_plus_linux.dart
@@ -46,7 +46,7 @@ class ShareLinux extends SharePlatform {
     String? subject,
     String? text,
     Rect? sharePositionOrigin,
-   ShareWithAppWindows? appName
+   ShareWithApp? appName
     }) {
     throw UnimplementedError('shareFiles() has not been implemented on Linux.');
   }