+++ packages/share_plus/share_plus_platform_interface/lib/method_channel/method_channel_share.dart
@@ -50,7 +50,7 @@ class MethodChannelShare extends SharePlatform {
     String? subject,
     String? text,
     Rect? sharePositionOrigin,
-    ShareWithAppWindows? appName
+    ShareWithApp? appName
   }) {
     assert(paths.isNotEmpty);
     assert(paths.every((element) => element.isNotEmpty));